py: create model registry client
Create a Python module for the model registry. We'll use gRPC to
communicate with a wrapper of the MLMD C++ server. For that, we'll use
the MLMD Python module because of its simplified interface, but also
define our own as we want to decouple from their implementation in the
future.

Signed-off-by: Isabella Basso do Amaral <idoamara@redhat.com>