Istio sample update (#76)
* Automation: Model-Registry-Isito-Deployment with ODH
This commit merges a script that will deploy Model-Registry with ODH then runs a rest api to conclude the script.
It intended for use with Openshift-CI.

Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>

* Revert to original script.
This commit reverts the script to deply basic model-registry back to it's original. Changes here are needed.

Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>

* Revert to original file
This commit reverts to the original file format as specifying opendatahub version is not required

Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>

* Creating specific rest script for tls
This commit introduces istio-tls- rest.sh script which is specific to the istio-tls deployment
It adds the certs/ to gitignore
It makes changes to the oci-model-registry-istio-tls-deploy.sh script to point to the correct rest.sh

Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>

* Update to gitactions
This commit is to allow the PR to pass the codcov tests

Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>

---------

Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>