This PR will add a Known Issues section to the README.md file. It also adds informaiton on a known issue regarding DB deployment and how to mitigate it when it happens
Signed-off-by: tonyxrmdavidson <tonyxrmdavidson@yahoo.co.uk>

Change 'Known Isues' to 'Tips'

This commit will change the title of the section of the README to Tips

Co-authored-by: Matteo Mortari <matteo.mortari@gmail.com>

Change Title to Generic Reference

This commit will change the title from Model-Registry-db mysql:8.3.0 to use a generic title which will reflect that pull image rate limiting is not confined to mysql:8.3.0

Co-authored-by: Matteo Mortari <matteo.mortari@gmail.com>

Change Description of Issue

This commit will change the description to use the model-registry-db container issue as an example to demonstrate image rate limiting.

Co-authored-by: Matteo Mortari <matteo.mortari@gmail.com>

Specify Error Description is an Example

This commit will expand the description of the error to show it is an example of a wider issue.

Co-authored-by: Matteo Mortari <matteo.mortari@gmail.com>