Add Tips section to README.md (#287)
Change 'Known Isues' to 'Tips'

This commit will change the title of the section of the README to Tips



Change Title to Generic Reference

This commit will change the title from Model-Registry-db mysql:8.3.0 to use a generic title which will reflect that pull image rate limiting is not confined to mysql:8.3.0



Change Description of Issue

This commit will change the description to use the model-registry-db container issue as an example to demonstrate image rate limiting.



Specify Error Description is an Example

This commit will expand the description of the error to show it is an example of a wider issue.

Co-authored-by: Matteo Mortari <matteo.mortari@gmail.com>