feat: add hint_server_address_port method with tests
Add static method to ModelRegistry client
that inspect server address protocol against port numbers,
issuing warnings for potentially mismatched HTTPS/HTTP configurations.
Method is called during ModelRegistry initialization to hint users to verify their setup.

For example, it took me a while to realize when copying a snippet from an example

```py
ModelRegistry("http://...",
    8080,
    author="myself",
    is_secure=False)
```

to another cluster where https:// had to be used,

```py
ModelRegistry("https://...",
    8080,
    author="myself")
```

as the error was a timeout error after several seconds,
and not an immediate connection error when invoked on that cluster.

Sometimes user might overlook this as it happened to me
so at this point emitting a warning could be beneficial.

This PR includes test suite covering different scenarios.

Signed-off-by: Matteo Mortari <matteo.mortari@gmail.com>