ci: ensure manual Fuzz Test use local built image
a race condition happens if the fuzz test starts
the :latest image is loaded in KinD
but a more recent image is available on ghcr.io
container registry, for example by merging and
a new commit on main.

typically other ci/GHA workflow pinned the IMG_VERSION
we are doing a tag containing a sha even tho it's not
necessary, unless we will ever tag and image on ghcr.io
with `:fuzz` but for consistency using the postfix sha

since the github event.after does not exists for
workflow_dispatch, using sha from the previously
existing step. so you might notice this is different
from other ci/GHA which are looking up the sha via the
PR event.

Signed-off-by: Matteo Mortari <matteo.mortari@gmail.com>

Co-authored-by: Alessio Pragliola <seth.pro@gmail.com>