fix user token handling (#1638)
Previously the user_token_envvar kwarg for ModelRegistry.__init__ was
given an explicit default string value, so passing in None or another
falsy value would override the default. With this change, passing a
falsy value results in the default being tried. This also errors when
the user passes a non-empty string value but the env var is not set.

This allows the async-job to automatically use the default value as
expected. Previously, we had to explicitly specify a value like this:
```
- name: MODEL_SYNC_REGISTRY_USER_TOKEN_ENVVAR
  value: KF_PIPELINES_SA_TOKEN_PATH
```
because failing to do so sets user_token_envvar to None.

Signed-off-by: Jon Burdo <jon@jonburdo.com>