ci: double-ensure E2E Fuzz just for Push to Main
for future-proofing the potential use of dev branches
we double-check the event that in the Python E2E testing
trigger the Fuzz testing, that is only when it's a
Push event to the Main branch (only)
keeping it consistent with the Step Name,
and double-checking the conditions for the run of this Step.

In the future, should a dev branch or stable branch
receive a PR to align `main` with these other branches,
the Event being Pull Request will not launch the Fuzz tests.

Please notice this condition is consistent with the current
workflow `on:` that includes indeed only push to main branch:

```yaml
name: Python workflows
on:
  push:
    branches:
      - "main"
...
```

Signed-off-by: Matteo Mortari <matteo.mortari@gmail.com>
Co-authored-by: Alessio Pragliola <seth.pro@gmail.com>